+++ e2e/test-env/src/client/RicosTestApp/RicosTestApp.tsx
@@ -22,6 +22,7 @@ import { themes } from '../consumersThemes/themes';
 import type { PaletteColors, DraftContent, SEOSettings } from 'wix-rich-content-common';
 import type { EditorState } from '@wix/draft-js';
 import { merge } from 'lodash';
+import { customStylesMock } from './customStylesMock';
 
 const VIEWER_ONLY = false;
 const onVideoSelected = (url: string, updateEntity) => {
@@ -36,35 +37,6 @@ const determinePalette = (paletteType: 'light' | 'dark', fallbackColor?: string)
 const setBackground = (palette: PaletteColors, disableContainer: boolean) =>
   !disableContainer && palette ? { backgroundColor: palette.bgColor } : {};
 const setForeground = (palette: PaletteColors) => (palette ? { color: palette.textColor } : {});
-const customStyles = [
-  'h1',
-  'h2',
-  'h3',
-  'h4',
-  'h5',
-  'h6',
-  'p',
-  'quote',
-  'link',
-  'hashtag',
-  'button',
-].reduce(
-  (prev, curr) => ({
-    ...prev,
-    [curr]: {
-      fontFamily: 'Times',
-      fontSize: '40px',
-      color: 'orange',
-      fontStyle: 'italic',
-      textDecoration: 'underline',
-      fontWeight: 'bold',
-      lineHeight: '40px',
-      minHeight: '40px',
-      borderColor: 'brown',
-    },
-  }),
-  {}
-);
 
 interface RicosTestAppProps {
   isMobile: boolean;
@@ -140,7 +112,7 @@ class RicosTestApp extends PureComponent<RicosTestAppProps> {
           palette,
           paletteConfig: { contentBgColor, settingsActionColor, focusActionColor },
           customStyles: useCustomStyles
-            ? customStyles
+            ? customStylesMock
             : useParagraphLineHeight
             ? { p: { lineHeight: '60px' } }
             : {},
@@ -187,7 +159,7 @@ class RicosTestApp extends PureComponent<RicosTestAppProps> {
           palette,
           paletteConfig: { contentBgColor, settingsActionColor, focusActionColor },
           customStyles: useCustomStyles
-            ? customStyles
+            ? customStylesMock
             : useParagraphLineHeight
             ? { p: { lineHeight: '60px' } }
             : {},